CS50-Web-Development-capstone/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
scrapy>=2.5,<3.0
papaparse>=5.3.0,<6.0.0
